requirements.txt
# CopyPasteEverything - Lightweight Clipboard Sync Tool
# Dependencies for Windows clipboard synchronization

customtkinter>=5.2.0
pillow>=10.0.0
websockets>=12.0
pyperclip>=1.8.2
pystray>=0.19.5
requests>=2.31.0
aiohttp>=3.9.0
cryptography>=41.0.0

